[autotest] Add script to run a BVT2 suite
The autotest proxy server is written such that it needs to shell out
to a separate tool in order to service each request that it receives.
Provide such a tool for the RunJob command, that can schedule a test
suite, block until results are available, and then return a summary of
the results.

BUG=chromium-os:25584,chromium-os:21118
TEST=./site_utils/run_suite.py -b x86-mario -i x86-mario-release/R19-1730.0.0-a1-b1630 -s test_suites/control.dummy

Change-Id: I52d4069be52f32babce40929a403e43ab051491a
Reviewed-on: https://gerrit.chromium.org/gerrit/15846
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>