hardware_trackpad: resume recording gesture files
There are 90+ gesture variations to record based on the config file.
It is desirable that a user could quit recording the recording task
any time, and resume recording gestures later. This CL adds a
command-line option which allows a user to resume recording gesture
files.

BUG=chromium-os:27556
TEST=Run the following procedure on a machine
$ cd /usr/local/autotest/site_tests/hardware_Trackpad
$ ./trackpad_record.py  # and then enter user name when prompted
$ ./trackpad_record.py -c # enter the same user name when prompted

Change-Id: Ic34872c6956d90024215199cbc4586e7c331c6e0
Reviewed-on: https://gerrit.chromium.org/gerrit/17681
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>