Add a WSGI conf for the Autotest Frontend.
Make it so that we can launch the Autotest frontend as a WSGI application.
This solves random permission problem mismatches when updating the dir since
the frontend runs as the owner of /usr/local/autotest now.
/usr/local/autotest/logs/rpc_logs no longer require ridiculous permissions to
log rpc information.
Releases will run smoother in general if the frontend is running as the users
code base checked out.

BUG=chromium-os:26414
TEST=Ran local instance via apache to ensure the Frontend components load.

Change-Id: Ic209e33a9b6d92cee68dcb01f2333b41c9330ccf
Reviewed-on: https://gerrit.chromium.org/gerrit/15903
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>