Add a "pool" option to Suite and Reimager
A pool is a set of machines with a specific label. The concept of pools is
needed to be able to ensure scheduling priority and to segrate our machine
population.

If no pool is set the previous behavior is followed. Jobs will be scheduled
against chromeos versions and reimager jobs will be scheduled against "board"

Modify VERSION_PREFIX to use a : instead of -.

Uodate run_suite to offer the optional --pool flag.
Update create_suite RPC interface to accept the pool option.
Update create_suite to assume all "suite names" are under
	test_suites/control.<suite name>
TEST=Unittest and development server end to end runs.
BUG=chromium-os:26491

Change-Id: I3677c956bee195e20ad63b4e034fa7c6611b1ce1
Reviewed-on: https://gerrit.chromium.org/gerrit/16130
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Frank Farzan <frankf@chromium.org>