Removed tests for -fstack-protector-strong and -D_FORTIFY_SOURCE=2.
The tests for these two hardening options were too unreliable. We were checking
for these two compiler options by inspecting some symbols on a stripped binary.
The problem is that the symbols may not be present even if the binary is built
with the hardened compiler options.

The plan is to implement these earlier (as part of the build process) so an
unsecure toolchain or unsecure packages can be detected. A proof-of-concept CL
is here: https://gerrit.chromium.org/gerrit/#change,15292.

BUG=chromium-os:24742
TEST=emerge-x86-zgb autotest-tests
./build_image --board=x86-zgb
./run_remote_tests --board=x86-zgb --remote=chromeos-zgb1
platform_ToolchainOptions

Change-Id: I3348df8db19e63ad679ac48afc904e29d912db24
Reviewed-on: https://gerrit.chromium.org/gerrit/16209
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jim Hebert <jimhebert@chromium.org>
Commit-Ready: asharif <asharif@chromium.org>
Tested-by: asharif <asharif@chromium.org>