[autotest] RPC retry logic shouldn't retry RPCs that can't succeed
RPC retry logic now sends a noop RPC upon failure of the chosen RPC.
If the noop succeeds, we know that autotest is functioning, but that
our RPC failed.  It's not worth retrying, because it's unlikely that
there was a transient failure; rather, there was probably something
wrong with the RPC we were trying to send.

BUG=chromium-os:26768
TEST=unit
TEST=run a job, stop the AFE.  It should continue retrying.
TEST=run a suite job with an incorrect board name.  It should not retry.
STATUS=Fixed

Change-Id: I30cdc5fb43881c86efc585cbc556a4bcca5ba0b8
Reviewed-on: https://gerrit.chromium.org/gerrit/16608
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>