autotest: wifi: Add Manager test for hidden networks
When we connect to a hidden network, ensure that "WiFi.HiddenSSID"
property is set to "True" by default on the service.  As a bonus
change, ensure that site_wlan_wait_state works correctly with a
timeout set to "0".  Also needed to fixup reporting of wait_service
to fix up the output keyvals.

BUG=chromium-os:26977
TEST=Run the test

Change-Id: I444110c5a9f731c58edd7a4acee1f0a9f00b5b6b
Reviewed-on: https://gerrit.chromium.org/gerrit/17036
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>