Check that the low-memory notification fires when memory is actually low.
The previous version of this test checked that the mechanism worked,
but not when.  This version additionally verifies that the mechanism
is actually doing what we think it's doing.

A future change will create pressure on the buffer cache by reading
lots of files, and check that the notification still fires as expected.

BUG=chromium-os:26853
TEST=tested manually and with run_remote_tests.

Change-Id: I3c7eaa6a9c898d95db87f40eded5de0270d5dad5
Reviewed-on: https://gerrit.chromium.org/gerrit/17087
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>