[autotest] Provide a more flexible dynamic_suite API
In practice, we will frequently be running control files from branches
in a ToT autotest environment.  In order to avoid being unable to
update and improve the dynamic_suite stuff, provide an API
that accepts **dargs.  This way, we can be sure that deprecated arguments
can be coped with, and that we don't have to worry about positional
arguments changing order.

BUG=chromium-os:26523
TEST=unittests
TEST=./server/autoserv test_suites/dev_harness

Change-Id: Iaf7a8844ea330346b654f9b4f51652b798d0ea9b
Reviewed-on: https://gerrit.chromium.org/gerrit/17123
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>