[autotest] Add platform_CloseOpenSimple
Existing servo tests are very much intended to be run at a developer's desk.
This CL takes a first cut at a test intended to be run with servo in the lab.
We eschew the ServoTest class, as it seems to be mostly designed to allow
command line args that come in via run_remote_tests.sh

We will handle the at-a-desk case later.

BUG=chromium-os:26666
TEST=This is a new test; run it.

Change-Id: Ib6438b375fcef4f8798349fa16c1da50756457a5
Reviewed-on: https://gerrit.chromium.org/gerrit/17975
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>