cell_tools.py: Don't pass None for timeout
There used to be a verifier argument to ConnecToCellular.
It was removed, but there was still a call to it that passed None as
the second argument.  This ended up passing None down as a timeout,
until it exploded with a TypeError when we added the current time and
None together.

This change removes that None.

BUG=chromium-os:25378
TEST=None, is one

Change-Id: Id55014e79ebfc05600732522b222c4128718ff8f
Reviewed-on: https://gerrit.chromium.org/gerrit/17237
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>