autotest (wifi): close race condition in WiFiRoaming.001SSIDSwitchBack
Before this patch, when the DUT gets DEAUTH from the "t2" AP, it
immediately rescans, and may find the "t2" AP before it shuts down.

This violates an assumption of the test -- that the DUT will
not see "t2" after we deconfig it. This leads (in some trials)
to the DUT attempting to reconnect to "t2". When this happens,
the DUT may take longer than allowed to re-connect to t1.
Alternatively, in some trials, flimflam eventually decides that
neither 't1' nor 't2' is viable.

Since the purpose of the test is just to acertain that the
connection manager will fall back to 't1' after 't2' fails,
modify the test to stop 't2' such that it does not send DEAUTH to
the DUT. Consequently, the DUT observes disconnection due to
beacon loss, and scans initiated by the DUT won't show 't2'
in scan results.

BUG=chromium-os:27148
TEST=WiFiRoaming.001SSIDSwitchBack (10 times with flimflam, all passed)

Change-Id: I393337c8a5b3355d70d34c0cb93a795171dadebc
Reviewed-on: https://gerrit.chromium.org/gerrit/17248
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: mukesh agrawal <quiche@chromium.org>