hardware_Trackpad: add verification log in test summary
For every criterion specified in the criteria list for every
functionality, there is a corresponding verification method
to examine whether a gesture file conforms to the criterion.
For example, a criterion for two-finger alt right click may
look like:

  'button':         ('Button Right', '==', 5),

The verification log of a failed test on a right-click gesture
file might look like:

  'button': {'Button Left': 2.0, 'Button Right': 3.0}

This CL collects all verification logs of distinct gesture sets
in the summary report

BUG=chromium-os:22569
TEST=No need to test this module until the web programs have been
checked in to capture gesture files.

Change-Id: If05a157bba8c7257a48496e2b4bcff00882bb46b
Reviewed-on: https://gerrit.chromium.org/gerrit/16245
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>