[autotest] Print log link for failed reimaging job
The dynamic_suite reimaging job gets kicked off by a mechanism that's
separate from the one we use for other jobs in a dynamic_suite.  As
such, it was not getting added to the suite job's keyval file, so we
were not able to have run_suite.py print log links for it upon
failure.

BUG=chromium-os:27281
TEST=./server/autoserv test_suites/dev_harness; ensure keyval file has try_new_image=xxx-username
STATUS=Fixed

Change-Id: I786785eded8bc4beb511fd8bbe944ac3712bd43c
Reviewed-on: https://gerrit.chromium.org/gerrit/17349
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>