Add a pathway to show main suite job errors.
If the suite job never runs a job and doesn't add in job keyvals
we end up with a failed SERVER_JOB that we ignore.

This is an attempt to expose that error with a link.

Example output:

SERVER_JOB   [ FAILED ]
SERVER_JOB     FAIL: global name '_vet_rmage_and_run_args' is not defined
@@@STEP_LINK@SERVER_JOB@http://chromeos-autotest.cbf/tko/retrieve_logs.cgi?job=/results/852-scottz/@@@

TEST=Ran against dev server.
BUG=chromium-os:27377

Change-Id: I86dfc1e794116565922847d73e97d8c482a787c1
Reviewed-on: https://gerrit.chromium.org/gerrit/17400
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>