[autotest] Fix frontend_wrappers_unittest.py flakiness
When we introduced the NOOP RPC check into the autotest RPC retrying logic,
the mocks used in the unit tests were only set up to expect _one_ attempt
to send an RPC.  If the jittered delay returned by the retry infrastrucutre
was long enough, this would be fine -- we'd timeout before trying to send a
second NOOP RPC.  If the delay was too short, we'd oversaturate the mock by
trying more than once.

Make try_noop_rpc() mocks expect to be called >= 1 times to avoid this issue.

BUG=chromium-os:27389
TEST=unit tests

Change-Id: I93679f0fe67e763210f6729b14efee94c74461ab
Reviewed-on: https://gerrit.chromium.org/gerrit/17414
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>