network_3GModemControl: Increase Enable timeout
The Y3300 seems to need around or a little over 10 seconds to enable and
register.  flimflam sometimes waits for it to register before setting
Powered=True, so increase the enable timeout long enough for it to enable
and register.

Also, add some logging so we can more easily see what the test is doing
without cross-referencing the source.

BUG=chromium-os:25378
TEST=is one, fails later (Connect)

Change-Id: Ifa154bf6bebcfd6abe140ccae1135e790d88d79b
Reviewed-on: https://gerrit.chromium.org/gerrit/17489
Reviewed-by: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>
Commit-Ready: ttuttle <ttuttle@chromium.org>