hardware_Trackpad: Use xev original window rather than mtplot
A recent change in X seemed to break xdpyinfo to display the window
in focus. Without a decent way to derive the window id when
launching mtplot, xev cannot be bound to mtplot to capture X events.
This CL switches back to xev's original pop up window to capture X
events as an initial work-around.

BUG=chromium-os:27646
TEST=Run trackpad autotest, it should execute normally.

Change-Id: Ia1db07d391894f43cd099038a7336c41bff9f887
Reviewed-on: https://gerrit.chromium.org/gerrit/17821
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>