hardware_Trackpad: set up gesture file path for regression test
The hardware_Trackpad autotest is usually run with the gesture file
path set in "/var/tmp/trackpad_test_data/test_files" on the machine.
The path is valid if a tester records gesture files on the machine
and then run autotest. For running regression test on pre-recorded
gesture files, it is needed to set up the gesture file path differently.

BUG=chromium-os:27649
TEST=Run trackpad autotest on alex with control.regression.
It should execute normally (no matter pass or fail).
(We only check in gesture files for alex at this time. Will check in
gesture files for other platforms later.)

Change-Id: Icbe40330e259caf0961148fdffd946e146563a76
Reviewed-on: https://gerrit.chromium.org/gerrit/17822
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>