factory: Add the camera/ALS performance tests
This CL adds the camera performance and ALS calibration test for AB
panels and the full machine. It uses the backend library in the cros
directory to do the performance evaluation for the camera module. The
AB only test will be run repetitively on a single separate Chromebook
for different AB panels in the factory floor. The full machine test
will be run once for each fully assembled DUT.

The code is largely based on the factory antenna test as both tests
shares a common workflow, with only the difference in the test
targets. The current implementation only reads the ambient ALS value
and saves it instead of doing a real calibration.

BUG=chrome-os-partner:6617,chrome-os-partner:6616
TEST=Run under factory autotest -> funcs that doesn't need fixture works

Change-Id: Ic1ad34083e2f7c444304e76bd68b1cd2dfaaf29d
Reviewed-on: https://gerrit.chromium.org/gerrit/17818
Reviewed-by: Choon Chng <choonc@google.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>