Allow groups of tests to be hidden/shown together.
By default tests that are part of a group are hidden.  The tests
for a group are all displayed on the RHS panel once at least one
test in the group becomes active.

Keyboard shortcuts are now managed entirely by the UI.  They are
automatically calculated based on the tests that are visible.
This should allow more easily having more tests.

BUG=chrome-os-partner:7928
TEST=create a test list with TestGroup tests, switch between tests
and observe RHS test list

Change-Id: If6d67388ed9533fbfb8a368d2b037783a742da3e
Reviewed-on: https://gerrit.chromium.org/gerrit/17925
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>