Modified Pkcs11InitUnderErrors to only run when Chaps is not enabled.
These tests are not really relevant for Chaps because they manipulate
Opencryptoki-specific files. They also don't work in the Chaps paradigm
because they attempt to work with a PKCS #11 token before cryptohome has
inserted any token. There is an issue logged (crosbug.com/27741) for the
new autotests that need to be written for Chaps.

BUG=chromium-os:21003
TEST=Run the platform_Pkcs11InitUnderErrors autotest with Chaps enabled
and disabled.

Change-Id: Ic6e716104c83ea834d176f42db56827a7c5004f0
Reviewed-on: https://gerrit.chromium.org/gerrit/18064
Commit-Ready: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>