site_utils: improve compiler checking
Look at the output of the preprocessor rather than producing an ELF in
a hardcoded location (like /tmp/a.out).  If the file is owned by someone
else, the tests will fail to detect things properly.  So avoid that by
not using a temp file at all.

BUG=None
TEST=running iotools.py by hand with diff compilers showed pie patch being applied as needed
TEST=running piglit.py by hand with diff compilers showed x86 being detected correctly

Change-Id: Ieb4541e2d79e90a08390ae110ab4cb681325e6b3
Reviewed-on: https://gerrit.chromium.org/gerrit/18388
Reviewed-by: asharif <asharif@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>