Start to fix up existing kernel suites to use new SUITE tag.
Tests in this are expected to be longer running than bvt
and/or buckets of many tests for which the results will
require a longer turnaround to triage and improve.

Test failures in these suites will not close the tree
(at this time) so they have EXPERIMENTAL=True.

Also added first two dynamic suite definition files
for kernel.

BUG=chromium-os:22487

TEST=local test with run_remote_tests.

Change-Id: I3ac0a4e79f4245c2c75bc08a7db33344acd1a280
Reviewed-on: https://gerrit.chromium.org/gerrit/18469
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>
Commit-Ready: Mike Truty <truty@chromium.org>