Update downloader and test_scheduler to utilize new staging.
We have hit a problem where the downloader code no longer properly
stages images for R19. This CL fixes that.

Add an option to downloader to stage images via the new devserver pathway.
IFF a board in the json config has 'release' in the name (Currently used to
signify TOT) and also has a download_devserver variable set to true. All other
boards with release in the name are skipped, these are currently downloaded
via the HWTest stage. All -RXX releases are processed in the old way still.

Update test_scheduler to get its control files from the dev server for all TOT
images. The new style puts the autotest files under /autotest/ instead of
/server/.

Fixed a bug where passing None to the dev server is interpreted as "None" and
causes issues because it is a valid string.

TEST=Local runs of downloader to ensure the pathway is being executed with the proper
variables.
BUG=chromium-os:28085

Change-Id: I9c5dc9e3a10ad02b9c6246beece11b8f31b0e5dd
Reviewed-on: https://gerrit.chromium.org/gerrit/18656
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>