hardware_Trackpad: add regression subset
We have checked in two sets of gesture files for regression test.
One set is short and the other is long. This CL adds the argument
to specify which test set to run with.

BUG=chromium-os:27462
TEST=Follow the following steps on alex:
$ cd /usr/local/autotest/site_tests/hardware_Trackpad
To run the short regression test which will take about 2 minutes
$ ../../bin/autotest control.regression
To run the long regression test which will take about 1 hour
$ ../../bin/autotest control.regression_long

Change-Id: I6d05fac199707047ac58e9e61f72406d36bc7af3
Reviewed-on: https://gerrit.chromium.org/gerrit/18941
Commit-Ready: Joseph Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>