Modify test_scheduler to verify control files, update daily_test_config.
Apparently we do not hang all the time when we trigger a download.
crosbug.com/28329
Add a check that looks at the response from the devserver, if it is a
bad response, we raise an error causing us to later retry scheduling the job.
Eventually, once these files are staged we will properly start the test.

Update the daily_test_config to reflect the chromeos_test_config updates.
This fixes daily config tests on TOT, they were not running.

TEST=Local runs verifying autotest job creation.
BUG=chromium-os:27213

Change-Id: I46c5312ad42b871d7328cdb19ca2a3575388e2e6
Reviewed-on: https://gerrit.chromium.org/gerrit/18986
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>