[autotest] Add de-duping suite scheduling logic for new scheduler
In the new scheduler, we'll rely on naming conventions to detect
duplicate jobs.  The DedupingScheduler class will schedule a job for a suite,
given the name, board and build to test -- unless it finds an existing
job.

For example, the existence of a job called:
  x86-mario-release/R18-1647.0.0-a1-b1567-test_suites/control.bvt

would prevent us from scheduling a 'bvt' suite for build
'x86-mario-release/R18-1647.0.0-a1-b1567' on board 'x86-mario'.

BUG=chromium-os:27586
TEST=new unit tests

Change-Id: Ida87804dd34600410fef33200881d30f939d9fd0
Reviewed-on: https://gerrit.chromium.org/gerrit/19188
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>