factory: Add image logging and a check for bad cam init for camperf test
The CL adds a functionality to automatically save the captured image to
an external USB disk when a fail is found during the AB panel test. It
also adds a check for a strange phenomenon that the captured image is
super small (160x120). The problem might be caused by the insufficient
bandwidth on the USB hub but we need to dig more to be sure.

BUG=chrome-os-partner:6617
TEST=Test it in the fixture -> works

Change-Id: I0bb9f77eaaa2104001359564e3e546aeb61503d1
Reviewed-on: https://gerrit.chromium.org/gerrit/19281
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>