[autotest] Add TriggerableJob and TimedTrigger to scheduler
The new scheduler will parse its config file and create a number
of TriggerableJob instances, which will be hung off of the
appropriate Trigger object.  So, if your job kicks off 'nightly',
the scheduler will attach the TriggerableJob for your job to the
Nightly object.  Then, every time the scheduler wakes up, it asks
each Trigger if it should be Fire()d.  If so, the scheduler
will Fire() the Trigger, which will Trigger() every TriggerableJob.
Note also that, if a given Trigger() call returns false, the Trigger
will throw away the associated TriggerableJob.  This allows for one-shot
Triggers.

BUG=chromium-os:27586
TEST=unit tests

Change-Id: Ibc892fd9afcd7a99594397000c6061224282b19f
Reviewed-on: https://gerrit.chromium.org/gerrit/19311
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>