Simplify test failure emails and send to Google group.
Setup a parallel email generator that does not use any config
file. Rather, send an email for every test failure on every
board to a Google Group. Test owners can then subscribe to
the group and setup a filter on either the test category
or the test naem.

Simplify test owner email registration format to include
the ability to specify groups as a regex and to supply an
optional test regex within a group regex to filter test failures.

BUG=chromium-os:28728
TEST=https://groups.google.com/a/google.com/forum/?fromgroups#!topicsearchin/chromeos-automated-test-failures/%22Autotest$20failure$20in$20pyauto_basic$20on$20x86-alex-r20$20(R20-2048.0.0-a1-b1855)%22

Change-Id: Ie115becd1b039e1dafc820ed8aed11ae82dae2b1
Reviewed-on: https://gerrit.chromium.org/gerrit/19346
Reviewed-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>
Commit-Ready: Mike Truty <truty@chromium.org>