Update site_wifitest to automatically discover cell hosts.
The wifitests currently rely on a configuration file to find out the
ip address of the router and the rspro. This CL updates this pathway
to check if the machine is running in the lab dns zone. If it is, the
lab host naming conventions are followed for both the router and the
rspro hosts that are required for this test to work.

TEST=Local autoserv instance showing that the correct names were used
when run on a lab machine and that the previous behavior is intake when running
on a non-lab machine.
BUG=chromium-os:28738

Change-Id: Ia2d0dbe24405ac2dd6181a6306fc32b950be1be4
Reviewed-on: https://gerrit.chromium.org/gerrit/19380
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>