[autotest] Re-name classes in the new suite_scheduler
While writing a docstring to explain the architecture of the new suite
scheduler, it became clear that I was running into name collisions.
So, I renamed things to be use more standard event-loop naming, and I
think it's a lot more understandable now.  Sorry for all the
confusion.

Also, added the beginnings of the command-line interface, with the
aforementioned big docstring at the top.

BUG=chromium-os:27586, chromium-os:28731
TEST=unit tests should still pass.

Change-Id: Ia8dd5149655ebcad86d27211d9c318bcd717e9b3
Reviewed-on: https://gerrit.chromium.org/gerrit/19423
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>