sound_infrastructure: Remove numid=<num> from controls files
The numid field for each of the alsa controls is a unique number
associated with the control.  The number changes based on how the
controls are enumerated by the driver. This enumeration can change as
drivers get refactored in the kernel causing test to fail
unnecessarily.

This CL removes the numid=<num> string as part of the comparison field
for the alsa controls in sound_infrastructure test.

BUG=chromium-os:28035
TEST=./run_remote_tests.sh --remote=<ip> sound_infrastructure
Passes on system w/ realtek ALC27x series codec

Change-Id: I271bde9432d26fb7af637b25b56f9ac6bae8580f
Reviewed-on: https://gerrit.chromium.org/gerrit/19472
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Taylor Hutt <thutt@chromium.org>
Commit-Ready: Todd Broch <tbroch@chromium.org>