factory: Parallelize MTF computation and add SN check for camperf test
The CL parallelize the time-consuming MTF computation in the camera
performance test library. For a conventional laptop with 2
hyper-threading cores, this provides about 2.2x of speedup and enables
us to compute MTF on all edges of the test pattern in reasonable time
(~10s). The benefit is that the result is now only subjective to the
image noise variation instead of the additional edge sampling variation,
which can be as high as ~0.3 in some cases.

Besides that, the CL also completes the serial number check fuctionality
in the factory test. The test will read a regular expression pattern
from the external USB disk and try to match the serial number with the
pattern. The design is flexible because the developer can modify the
pattern according to the factory format on site without the need to
modify the code.

Finally, the CL also refactored the code a little bit for better clarity
and add a minimum MTF check for the image sharpness test.

BUG=chrome-os-partner:6617
TEST=Test it with real images -> works

Change-Id: I884a962357886265c130355f840bb7eb6e98f98d
Reviewed-on: https://gerrit.chromium.org/gerrit/19661
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>