Kludge to allow for the test_scheduler to find milestones in -release targets.
Since we are staging everything in board-release we now have this issue where
looking for the latest build for a particular milestone requires us to pass
that parameter, otherwise we always get the TOT build.

We are close to replacing this code. This is a bruteforce hack. Since we use
mkdir for locking (sigh) we do not kick off any extra jobs here if no builds
are rolling.

TEST=Ran test run and verified that jobs are scheduled.
BUG=chromium-os:29087

Change-Id: Iaad4efe2f5b46b27409fd3378912ceb8d6cf85df
Reviewed-on: https://gerrit.chromium.org/gerrit/19780
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>