[autotest] End job in ERROR if there are _no_ appropriate hosts
Currently, if the caller asks the dynamic_suite code to pre-check for enough
hosts, we will WARN if there are too few devices to run the test suite.  Even
if there are none at all.  We've decided to escalate the case where all hosts
are down to be an ERROR, so that the HWTest stage will fail fatally.

BUG=chromium-os:29163
TEST=unit
TEST=run_suite.py with a valid board that has no hosts.
STATUS=Fixed

Change-Id: I310b2a3bc98572c619207a1e4013ea9e56994539
Reviewed-on: https://gerrit.chromium.org/gerrit/19896
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>