[autotest] Enable forcing an event to fire and schedule suites
We need to ability to make the scheduler force an event, so we can backfill
test results, if need be.  This is also handy for testing stuff.

Also, fix a bug in comparing build branch names to branch specifications
encoded in Tasks.  Before, every Task would run on every release branch,
even if the Task said it should only run on factory and/or firmware.

BUG=chromium-os:27586
TEST=unit
TEST=./suite_scheduler.py -f suite_scheduler_test.ini -e nightly -i x86-mario-release/R20-2130.0.0-a1-b1905

Change-Id: I58217c358d8b85faf49a940d01bca184dc8590e0
Reviewed-on: https://gerrit.chromium.org/gerrit/20124
Commit-Ready: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>