autotest: graphics_GpuReset initial checkin.
The purpose of this test is to exercise the GPU failure path.
We craft an erroneous GPU command packet and send it to the GPU,
and wait for a udev event notifying us of a GPU hang.
If the event doesn't come back, the test fails.

TEST=Ran on lumpy and tested different failure modes.

Change-Id: Ibdb8e638f5d218e48905ad7813116bb16f8f040c
Reviewed-on: https://gerrit.chromium.org/gerrit/20294
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>