Add support for gs_offloader to check if a job has completed.
Instead of looking at timestamps to decide what to offload, ask the AFE
if the job has completed. If it has, offload it.

Introduce a simple is_job_complete script that can be used for one off scripts
and is used in gs_offloader to verify a job is complete.

Add a ditch RSYNC option that can be enabled in the script itself. This is handy
when the HD fills up.

TEST=Code that has been running live for quite some time that has been cleaned
up and added to review. :(
BUG=chromium-os:24540

Change-Id: I7eb4f0975a9c6330c5218367955e016ec461e4ac
Reviewed-on: https://gerrit.chromium.org/gerrit/20378
Commit-Ready: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Tested-by: Scott Zawalski <scottz@chromium.org>