Make the servo object a property of the host.
This adds a 'servo' member to instances of SiteHost, when the
target hostname has an associated "-servo" hostname as well.

Additionally, in order to support run_remote_tests, an optional
parameter to the host initialize() method can be used to specify
that servod should be started on the local host if no "-servo"
host can be found.

Not all test control files are updated to use the new scheme,
so class ServoTest retains the ability to create and start
servod itself, when necessary.

BUG=chromium-os:29209
TEST=run all the changed tests using a locally attached Servo
TEST=same test with servo debug header detached to force failure

Change-Id: Ibbd521ee5f90450fd5fa30629a020ba3fd45937d
Reviewed-on: https://gerrit.chromium.org/gerrit/20657
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>